backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pydantic==2.10.3
pydantic-settings==2.6.1
pandas==2.2.3
openpyxl==3.1.5
httpx==0.28.1
resend==4.0.0
python-dotenv==1.0.1
